uint32_t conversations. save / load version 0.   - replaced all ulong's to uint32_t. It is need to good synchronization     between 32-bit and 64-bit platforms.   - progress in creating version 0 of loading / saving competition format.

+++ src/core/base/rstring.cpp
@@ -58,10 +58,10 @@ rfc::String rfc::String::toString(long number, long signsCnt)
  * The problem is fgets doesn't works correctly with
  * binary opened files. ( WHY ??? ).
  */
-void rfc::String::fgets(char *data, const ulong maxLen, FILE * &fileIn)
+void rfc::String::fgets(char *data, const uint32_t maxLen, FILE * &fileIn)
 {
 	char ch;
-	ulong len = 0;
+	uint32_t len = 0;
 
 	while (len < maxLen) {
 		ch = fgetc(fileIn);
